rag-api/requirements.txt
# RAG API Dependencies
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0

# LlamaIndex Core
llama-index-core>=0.14.0
llama-index-llms-openai>=0.4.0
llama-index-embeddings-huggingface>=0.5.0
